requirements.txt
numpy==2.4.1
opencv_python==4.12.0.88
opendatasets==0.1.22
Pillow==12.1.0
torch==2.7.1+cu118
torchvision==0.22.1+cu118
